requirements.txt
asn1crypto==1.5.1
Babel==2.9.1
bcrypt==4.0.1
bleach==5.0.1
blinker==1.5
cbor2==5.4.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
email-validator==1.3.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gitdb==4.0.9
GitPython==3.1.18
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mkdocs-material-extensions==1.1
mongoengine==0.23.1
openpyxl==3.0.9
passlib==1.7.4
phonenumberslite==8.13.0
pycparser==2.21
pydantic==1.10.2
pymongo==3.12.1
pyOpenSSL==22.1.0
PyQRCode==1.2.1
python-dotenv==0.21.0
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
six==1.16.0
smmap==5.0.0
speaklater==1.3
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
typing_extensions==4.4.0
urllib3==1.26.12
webauthn==1.6.0
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
